arg_reader: Fix errors when reading last part of contiguous segments from file
The contiguous range might fail because it does not have all
the characters, so we fall back and scan the range as not contiguous

A basic_scan_file_buffer, when it has buffering, it can become
discontiguous while moving forward when executing
ranges::next(range.begin(), ranges::distance(crange.begin(), it)).
E.g. the contiguous buffer might have 4.1234, but the full number is 4.12345
For that reason, we fall back and scan the range as not contiguous

Also, it might not read all the digits because it does not have them all
E.g. if it has 1.1234e, but it needed 1.1234e-01, it will disregard the e.
For that reason, we fall back and scan the range as not contiguous