win32/default: Enforce only 1 successfully putback guarantee
https://en.cppreference.com/w/c/io/ungetc states that only 1 byte
is guaranteed to be put back. Even if linux/mac support 4KB,
windows and other platforms are very strict with 1 character,
and they can fail if you try 2-3-4 ungetc.

To that extent, no more characters than needed are ever consumed.

Also, when parsing from a file, if whitespace has been consumed,
it remains consumed, it's not put back, similar to fscanf.
https://pubs.opengroup.org/onlinepubs/000095399/functions/fscanf.html